config: fix logic audit findings (strict mode)
Critical fixes:
- file_ops: Use checked_mul to prevent u64 overflow in file size parsing
- schema: Replace expect() with unwrap_or_else + error logging
- patcher: Replace expect() in schema cache with safe fallback

Warning fixes:
- retry: Fix f64 precision loss (u64::MAX as f64) and i32 truncation
- metrics: Fix f64 precision loss in warning threshold
- text/search: Fix usize/u64 portability with try_into
- toml_ops: Propagate fsync errors instead of silently ignoring
- save: Collapse nested if to satisfy clippy

+++ src/config/types/policies/metrics.rs
@@ -107,7 +107,7 @@ impl MetricsPolicy {
         if !self.warning_multiplier.is_finite() || self.warning_multiplier < 0.0 {
             return target_ms;
         }
-        (target_ms as f64 * self.warning_multiplier).clamp(0.0, u64::MAX as f64) as u64
+        (target_ms as f64 * self.warning_multiplier).clamp(0.0, f64::MAX) as u64
     }
 
     /// Check if a duration exceeds the warning threshold for hotkey->clipboard
