+++ src/config/types/policies/retry.rs
@@ -110,8 +110,8 @@ impl RetryPolicy {
         }
 
         let backoff_ms = (self.initial_backoff_ms as f64
-            * self.backoff_multiplier.powi(attempt as i32))
-        .clamp(0.0, u64::MAX as f64) as u64;
+            * self.backoff_multiplier.powi(attempt.try_into().unwrap_or(i32::MAX)))
+        .clamp(0.0, f64::MAX) as u64;
         let capped = backoff_ms.min(self.max_backoff_ms);
         std::time::Duration::from_millis(capped.max(self.initial_backoff_ms))
     }
